Refactor reason code logic into sub service
